requirements.txt
tensorflow==2.16.1
Flask
numpy
pillow
flask-cors